+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 - Cmdlets now return a suitable `PSCustomObject` with categorised output
 - Add support for retrieving devices with a status other than 'OK' (**Windows 10/Server 2016 only**)
+- Add support for checking for kernel & service profile crash dumps (`LocalSystem`, `LocalService` & `NetworkService`)
 - Major clean-up of code (stylistic improvements, stop using `Write-Host`, etc...)
 
 ## v0.1